Reduced code-size + STVP programming option added
Reduced code-size by removing all bit-fields and unions. This saved over
500 bytes of Flash memory.

+++ src/pid.c
@@ -94,9 +94,9 @@ void pid_ctrl(int16_t yk, int16_t *uk, uint16_t tset)
     //                                      Ti           Ts
     //
     //-----------------------------------------------------------------------------
-    pp   = (uint32_t)kc * (yk_1 - yk);     // Kc.(y[k-1]-y[k])
-    pp  += ki * (tset - yk);               // (Kc.Ts/Ti).e[k]
-    pp  += kd * ((yk_1 << 1) - yk - yk_2); // (Kc.Td/Ts).(2.y[k-1]-y[k]-y[k-2])
+    pp   = (uint32_t)kc * (yk_1 - yk);               // Kc.(y[k-1]-y[k])
+    pp  += (uint32_t)ki * (tset - yk);               // (Kc.Ts/Ti).e[k]
+    pp  += (uint32_t)kd * ((yk_1 << 1) - yk - yk_2); // (Kc.Td/Ts).(2.y[k-1]-y[k]-y[k-2])
     *uk += (int16_t)pp;
     // limit u[k] to GMA_HLIM and GMA_LLIM
     if (*uk > GMA_HLIM)      *uk = GMA_HLIM;
